[SUBMARINE-236]  Move the GitUtils from org.apache.submarine.server to org.apache.submarine.database.utils.
