+++ submarine-workbench/workbench-server/src/main/java/org/apache/submarine/database/utils/GitUtils.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.submarine.server;
+package org.apache.submarine.database.utils;
 
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.ListBranchCommand;
